Form mostly refactored. Lacking employer address if distinct.

+++ db/migrate/20151104013924_create_claims.rb
@@ -6,7 +6,7 @@ class CreateClaims < ActiveRecord::Migration
       t.string :award
       t.string :lost_wages
       t.integer :total_hours
-      t.integer :hourly_pay
+      t.decimal :hourly_pay
       t.references :employer, index: true, foreign_key: true
       t.references :user, index: true, foreign_key: true
       t.date :employment_began_on
